Use two enums outside of the primary class definition
The enums correspond to relational operators that compare a single
value to zero (`OpcodeCompareToZero`) and relational operators that
compare two values to each other (`OpcodeCompare`).
`OpcodeCompareToZero` is the type of the `opcode` attribute of
`RelationalOperatorReplacementMutator`.

Creating enum values that correspond to all twelve relational
operators will be useful when writing the class's `MethodVisitor()`.